ATR-180: unified code generation utils
