+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/DacUiAttributes/DacUiAttributesFix.cs
@@ -4,8 +4,8 @@ using System.Composition;
 using System.Threading;
 using System.Threading.Tasks;
 
+using Acuminator.Utilities.Roslyn.CodeGeneration;
 using Acuminator.Utilities.Roslyn.Semantic;
-using Acuminator.Utilities.Roslyn.Syntax;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeActions;
