--- src/Acuminator/Acuminator.Utilities/Roslyn/Syntax/SyntaxNodeGenerationUtils.cs
@@ -1,40 +1,0 @@
-﻿#nullable enable
-
-using Acuminator.Utilities.Common;
-
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Simplification;
-
-namespace Acuminator.Utilities.Roslyn.Syntax
-{
-	public static class SyntaxNodeGenerationUtils
-	{
-		/// <summary>
-		/// Create attribute list of the supplied type
-		/// </summary>
-		/// <param name="type"></param>
-		/// <returns></returns>
-		public static AttributeListSyntax GetAttributeList(this INamedTypeSymbol type, AttributeArgumentListSyntax? argumentList = null)
-		{
-			type.ThrowOnNull();
-
-			var node = SyntaxFactory.Attribute(
-				SyntaxFactory.IdentifierName(
-					type.Name))
-				.WithAdditionalAnnotations(Simplifier.Annotation);
-
-			if (argumentList != null)
-			{
-				node = node.WithArgumentList(argumentList);
-			}
-
-			var list = SyntaxFactory.AttributeList(
-				SyntaxFactory.SingletonSeparatedList(
-					node));
-
-			return list;
-		}
-	}
-}
